Take out shared behavior for user images and put it in Eliza
