+++ kifi-backend/modules/shoebox/app/com/keepit/common/store/S3ImageStore.scala
@@ -33,8 +33,7 @@ import scala.util.{ Failure, Success, Try }
 import play.api.Play.current
 
 @ImplementedBy(classOf[S3ImageStoreImpl])
-trait S3ImageStore {
-  def config: S3ImageConfig
+trait S3ImageStore extends S3ExternalIdImageStore {
 
   def getPictureUrl(width: Int, user: User): Future[String]
   def getPictureUrl(width: Option[Int], user: User, picName: String): Future[String]
@@ -49,26 +48,7 @@ trait S3ImageStore {
   // Returns Some(urlOfUserPicture) or None
   def copyTempFileToUserPic(userId: Id[User], userExtId: ExternalId[User], token: String, cropAttributes: Option[ImageCropAttributes]): Option[String]
 
-  def avatarUrlByExternalId(w: Option[Int], userId: ExternalId[User], picName: String, protocolDefault: Option[String] = None): String = {
-    val size = S3UserPictureConfig.ImageSizes.find(size => w.exists(size >= _)).map(_.toString).getOrElse(S3UserPictureConfig.OriginalImageSize)
-    val uri = URI.parse(s"${config.cdnBase}/${keyByExternalId(size, userId, picName)}").get
-    URI(uri.scheme orElse protocolDefault, uri.userInfo, uri.host, uri.port, uri.path, uri.query, uri.fragment).toString
-  }
 
-  def avatarUrlByUser(user: User): String = avatarUrlByUser(BasicUser.fromUser(user))
-
-  def avatarUrlByUser(user: BasicUser): String =
-    avatarUrlByExternalId(Some(200), user.externalId, user.pictureName, Some("https"))
-
-  def keyByExternalId(size: String, userId: ExternalId[User], picName: String): String = {
-    val pic = if (picName.endsWith(".jpg")) picName else s"$picName.jpg"
-    s"users/$userId/pics/$size/$pic"
-  }
-
-  def tempPath(token: String): String = {
-    val pic = if (token.endsWith(".jpg")) token else s"$token.jpg"
-    s"temp/user/pics/$pic"
-  }
 }
 
 @Singleton
